pyproject.toml
[project]
name = "xianyuautoagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai==1.65.5",
    "websockets==13.1",
    "loguru==0.7.3",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "chromadb>=0.4.0,<1.0",
    "langgraph>=0.2.0,<1.0",
    "jieba>=0.42.1,<1.0",
]
